+++ zcash_primitives/src/sapling/keys.rs
@@ -292,6 +292,16 @@ impl DiversifiableFullViewingKey {
         zip32::sapling_default_address(&self.fvk, &self.dk)
     }
 
+    /// Returns the internal address corresponding to the smallest valid diversifier index,
+    /// along with that index.
+    ///
+    /// This address **MUST NOT** be encoded and exposed to end users. User interfaces
+    /// should instead mark these notes as "change notes" or "internal wallet operations".
+    pub fn change_address(&self) -> (zip32::DiversifierIndex, PaymentAddress) {
+        let internal_dfvk = self.derive_internal();
+        zip32::sapling_default_address(&internal_dfvk.fvk, &internal_dfvk.dk)
+    }
+
     /// Attempts to decrypt the given address's diversifier with this full viewing key.
     ///
     /// This method extracts the diversifier from the given address and decrypts it as a